delte DokumentService.kt til 3 klasser, 1. DokumentGenereringService.kt for å generere brev, 2. DokumentDistribueringService.kt for å distribuere brev og 3. DokumentService.kt for å sende brev og andre operasjoner (#3252)
