+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/DokumentService.kt
@@ -1,13 +1,10 @@
 package no.nav.familie.ba.sak.kjerne.brev
 
-import io.micrometer.core.instrument.Counter
-import io.micrometer.core.instrument.Metrics
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.FunksjonellFeil
 import no.nav.familie.ba.sak.config.RolleConfig
 import no.nav.familie.ba.sak.config.TaskRepositoryWrapper
 import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.DEFAULT_JOURNALFØRENDE_ENHET
-import no.nav.familie.ba.sak.integrasjoner.familieintegrasjoner.IntegrasjonClient
 import no.nav.familie.ba.sak.integrasjoner.journalføring.UtgåendeJournalføringService
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.DbJournalpost
 import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.DbJournalpostType
@@ -15,18 +12,11 @@ import no.nav.familie.ba.sak.integrasjoner.journalføring.domene.JournalføringR
 import no.nav.familie.ba.sak.integrasjoner.organisasjon.OrganisasjonService
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
-import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.behandling.settpåvent.SettPåVentService
 import no.nav.familie.ba.sak.kjerne.brev.domene.ManueltBrevRequest
 import no.nav.familie.ba.sak.kjerne.brev.domene.erTilInstitusjon
-import no.nav.familie.ba.sak.kjerne.brev.domene.maler.Brev
-import no.nav.familie.ba.sak.kjerne.brev.domene.maler.Brevmal
-import no.nav.familie.ba.sak.kjerne.brev.domene.tilBrev
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRepository
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
-import no.nav.familie.ba.sak.kjerne.logg.LoggService
 import no.nav.familie.ba.sak.kjerne.steg.BehandlerRolle
-import no.nav.familie.ba.sak.kjerne.steg.StegType
 import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.VilkårsvurderingForNyBehandlingService
 import no.nav.familie.ba.sak.kjerne.vedtak.Vedtak
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
@@ -35,15 +25,12 @@ import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.leggTilBlankAnnenVurdering
 import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
 import no.nav.familie.ba.sak.task.DistribuerDokumentDTO
 import no.nav.familie.ba.sak.task.DistribuerDokumentTask
-import no.nav.familie.ba.sak.task.DistribuerDødsfallDokumentPåFagsakTask
-import no.nav.familie.http.client.RessursException
 import no.nav.familie.kontrakter.felles.BrukerIdType
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.dokarkiv.AvsenderMottaker
 import no.nav.familie.kontrakter.felles.dokarkiv.v2.Førsteside
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDate
@@ -51,13 +38,8 @@ import java.util.Properties
 
 @Service
 class DokumentService(
-    private val persongrunnlagService: PersongrunnlagService,
-    private val integrasjonClient: IntegrasjonClient,
-    private val loggService: LoggService,
     private val journalføringRepository: JournalføringRepository,
     private val taskRepository: TaskRepositoryWrapper,
-    private val brevKlient: BrevKlient,
-    private val brevService: BrevService,
     private val vilkårsvurderingService: VilkårsvurderingService,
     private val vilkårsvurderingForNyBehandlingService: VilkårsvurderingForNyBehandlingService,
     private val rolleConfig: RolleConfig,
@@ -65,27 +47,11 @@ class DokumentService(
     private val utgåendeJournalføringService: UtgåendeJournalføringService,
     private val fagsakRepository: FagsakRepository,
     private val organisasjonService: OrganisasjonService,
-    private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService
+    private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService,
+    private val dokumentGenereringService: DokumentGenereringService
 ) {
 
     val logger: Logger = LoggerFactory.getLogger(this::class.java)
-
-    private val antallBrevSendt: Map<Brevmal, Counter> = mutableListOf<Brevmal>().plus(Brevmal.values()).associateWith {
-        Metrics.counter(
-            "brev.sendt",
-            "brevtype",
-            it.visningsTekst
-        )
-    }
-
-    private val antallBrevIkkeDistribuertUkjentAndresse: Map<Brevmal, Counter> =
-        mutableListOf<Brevmal>().plus(Brevmal.values()).associateWith {
-            Metrics.counter(
-                "brev.ikke.sendt.ukjent.andresse",
-                "brevtype",
-                it.visningsTekst
-            )
-        }
 
     fun hentBrevForVedtak(vedtak: Vedtak): Ressurs<ByteArray> {
         if (SikkerhetContext.hentHøyesteRolletilgangForInnloggetBruker(rolleConfig) == BehandlerRolle.VEILEDER && vedtak.stønadBrevPdF == null) {
@@ -97,66 +63,13 @@ class DokumentService(
         }
     }
 
-    fun genererBrevForVedtak(vedtak: Vedtak): ByteArray {
-        try {
-            if (!vedtak.behandling.skalBehandlesAutomatisk && vedtak.behandling.steg > StegType.BESLUTTE_VEDTAK) {
-                throw FunksjonellFeil("Ikke tillatt å generere brev etter at behandlingen er sendt fra beslutter")
-            }
-
-            val målform = persongrunnlagService.hentSøkersMålform(vedtak.behandling.id)
-            val vedtaksbrev =
-                when (vedtak.behandling.opprettetÅrsak) {
-                    BehandlingÅrsak.DØDSFALL_BRUKER -> brevService.hentDødsfallbrevData(vedtak)
-                    BehandlingÅrsak.KORREKSJON_VEDTAKSBREV -> brevService.hentKorreksjonbrevData(vedtak)
-                    else -> brevService.hentVedtaksbrevData(vedtak)
-                }
-            return brevKlient.genererBrev(målform.tilSanityFormat(), vedtaksbrev)
-        } catch (feil: Throwable) {
-            if (feil is FunksjonellFeil) throw feil
-
-            throw Feil(
-                message = "Klarte ikke generere vedtaksbrev på behandling ${vedtak.behandling}: ${feil.message}",
-                frontendFeilmelding = "Det har skjedd en feil, og brevet er ikke sendt. Prøv igjen, og ta kontakt med brukerstøtte hvis problemet vedvarer.",
-                httpStatus = HttpStatus.INTERNAL_SERVER_ERROR,
-                throwable = feil
-            )
-        }
-    }
-
-    fun genererManueltBrev(
-        manueltBrevRequest: ManueltBrevRequest,
-        erForhåndsvisning: Boolean = false
-    ): ByteArray {
-        Result.runCatching {
-            val brev: Brev = manueltBrevRequest.tilBrev { integrasjonClient.hentLandkoderISO2() }
-            return brevKlient.genererBrev(
-                målform = manueltBrevRequest.mottakerMålform.tilSanityFormat(),
-                brev = brev
-            )
-        }.fold(
-            onSuccess = { it },
-            onFailure = {
-                if (it is Feil) {
-                    throw it
-                } else {
-                    throw Feil(
-                        message = "Klarte ikke generere brev for ${manueltBrevRequest.brevmal}. ${it.message}",
-                        frontendFeilmelding = "${if (erForhåndsvisning) "Det har skjedd en feil" else "Det har skjedd en feil, og brevet er ikke sendt"}. Prøv igjen, og ta kontakt med brukerstøtte hvis problemet vedvarer.",
-                        httpStatus = HttpStatus.INTERNAL_SERVER_ERROR,
-                        throwable = it
-                    )
-                }
-            }
-        )
-    }
-
     @Transactional
     fun sendManueltBrev(
         manueltBrevRequest: ManueltBrevRequest,
         behandling: Behandling? = null,
         fagsakId: Long
     ) {
-        val generertBrev = genererManueltBrev(manueltBrevRequest)
+        val generertBrev = dokumentGenereringService.genererManueltBrev(manueltBrevRequest)
 
         val førsteside = if (manueltBrevRequest.brevmal.skalGenerereForside()) {
             Førsteside(
@@ -191,10 +104,7 @@ class DokumentService(
             )
         }
 
-        if (
-            behandling != null &&
-            manueltBrevRequest.brevmal.førerTilOpplysningsplikt()
-        ) {
+        if (behandling != null && manueltBrevRequest.brevmal.førerTilOpplysningsplikt()) {
             leggTilOpplysningspliktIVilkårsvurdering(behandling)
         }
 
@@ -265,86 +175,7 @@ class DokumentService(
             .leggTilBlankAnnenVurdering(AnnenVurderingType.OPPLYSNINGSPLIKT)
     }
 
-    fun prøvDistribuerBrevOgLoggHendelse(
-        journalpostId: String,
-        behandlingId: Long?,
-        loggBehandlerRolle: BehandlerRolle,
-        brevmal: Brevmal
-    ) = try {
-        distribuerBrevOgLoggHendlese(journalpostId, behandlingId, brevmal, loggBehandlerRolle)
-    } catch (ressursException: RessursException) {
-        logger.info("Klarte ikke å distribuere brev til journalpost $journalpostId på behandling $behandlingId. Httpstatus ${ressursException.httpStatus}")
-        secureLogger.info(
-            "Klarte ikke å distribuere brev til journalpost $journalpostId på behandling $behandlingId.\n" +
-                "Httpstatus: ${ressursException.httpStatus}\n" +
-                "Melding: ${ressursException.cause?.message}"
-        )
-
-        when {
-            mottakerErIkkeDigitalOgHarUkjentAdresse(ressursException) && behandlingId != null ->
-                loggBrevIkkeDistribuertUkjentAdresse(journalpostId, behandlingId, brevmal)
-
-            mottakerErDødUtenDødsboadresse(ressursException) && behandlingId != null ->
-                håndterMottakerDødIngenAdressePåBehandling(journalpostId, brevmal, behandlingId)
-
-            dokumentetErAlleredeDistribuert(ressursException) ->
-                logger.warn(alleredeDistribuertMelding(journalpostId, behandlingId))
-
-            else -> throw ressursException
-        }
-    }
-
-    internal fun håndterMottakerDødIngenAdressePåBehandling(
-        journalpostId: String,
-        brevmal: Brevmal,
-        behandlingId: Long
-    ) {
-        val task = DistribuerDødsfallDokumentPåFagsakTask.opprettTask(journalpostId = journalpostId, brevmal = brevmal)
-        taskRepository.save(task)
-        logger.info("Klarte ikke å distribuere brev for journalpostId $journalpostId på behandling $behandlingId. Bruker har ukjent dødsboadresse.")
-        loggService.opprettBrevIkkeDistribuertUkjentDødsboadresseLogg(
-            behandlingId = behandlingId,
-            brevnavn = brevmal.visningsTekst
-        )
-    }
-
-    internal fun loggBrevIkkeDistribuertUkjentAdresse(
-        journalpostId: String,
-        behandlingId: Long,
-        brevMal: Brevmal
-    ) {
-        logger.info("Klarte ikke å distribuere brev for journalpostId $journalpostId på behandling $behandlingId. Bruker har ukjent adresse.")
-        loggService.opprettBrevIkkeDistribuertUkjentAdresseLogg(
-            behandlingId = behandlingId,
-            brevnavn = brevMal.visningsTekst
-        )
-        antallBrevIkkeDistribuertUkjentAndresse[brevMal]?.increment()
-    }
-
-    private fun distribuerBrevOgLoggHendlese(
-        journalpostId: String,
-        behandlingId: Long?,
-        brevMal: Brevmal,
-        loggBehandlerRolle: BehandlerRolle
-    ) {
-        integrasjonClient.distribuerBrev(journalpostId = journalpostId, distribusjonstype = brevMal.distribusjonstype)
-
-        if (behandlingId != null) {
-            loggService.opprettDistribuertBrevLogg(
-                behandlingId = behandlingId,
-                tekst = brevMal.visningsTekst,
-                rolle = loggBehandlerRolle
-            )
-        }
-
-        antallBrevSendt[brevMal]?.increment()
-    }
-
     companion object {
         val secureLogger = LoggerFactory.getLogger("secureLogger")
-
-        fun alleredeDistribuertMelding(journalpostId: String, behandlingId: Long?) =
-            "Journalpost med Id=$journalpostId er allerede distiribuert. Hopper over distribuering." +
-                if (behandlingId != null) " BehandlingId=$behandlingId." else ""
     }
 }
