+++ src/main/kotlin/no/nav/familie/ba/sak/task/DistribuerVedtaksbrevTilVergeTask.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ba.sak.task
 
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
-import no.nav.familie.ba.sak.kjerne.brev.DokumentService
+import no.nav.familie.ba.sak.kjerne.brev.DokumentDistribueringService
 import no.nav.familie.ba.sak.kjerne.brev.hentBrevmal
 import no.nav.familie.ba.sak.kjerne.steg.BehandlerRolle
 import no.nav.familie.kontrakter.felles.objectMapper
@@ -12,18 +12,23 @@ import org.springframework.stereotype.Service
 import java.util.Properties
 
 @Service
-@TaskStepBeskrivelse(taskStepType = DistribuerVedtaksbrevTilVergeTask.TASK_STEP_TYPE, beskrivelse = "Send vedtaksbrev til verge til Dokdist", maxAntallFeil = 3)
+@TaskStepBeskrivelse(
+    taskStepType = DistribuerVedtaksbrevTilVergeTask.TASK_STEP_TYPE,
+    beskrivelse = "Send vedtaksbrev til verge til Dokdist",
+    maxAntallFeil = 3
+)
 class DistribuerVedtaksbrevTilVergeTask(
     private val behandlingHentOgPersisterService: BehandlingHentOgPersisterService,
-    private val dokumentService: DokumentService
+    private val dokumentDistribueringService: DokumentDistribueringService
 ) : AsyncTaskStep {
 
     override fun doTask(task: Task) {
-        val distribuerVedtaksbrevDTO = objectMapper.readValue(task.payload, DistribuerVedtaksbrevTilVergeDTO::class.java)
+        val distribuerVedtaksbrevDTO =
+            objectMapper.readValue(task.payload, DistribuerVedtaksbrevTilVergeDTO::class.java)
 
         val behandling = behandlingHentOgPersisterService.hent(distribuerVedtaksbrevDTO.behandlingId)
 
-        dokumentService.prøvDistribuerBrevOgLoggHendelse(
+        dokumentDistribueringService.prøvDistribuerBrevOgLoggHendelse(
             journalpostId = distribuerVedtaksbrevDTO.journalpostId,
             behandlingId = distribuerVedtaksbrevDTO.behandlingId,
             loggBehandlerRolle = BehandlerRolle.SYSTEM,
