+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/brev/DokumentServiceEnhetstest.kt
@@ -16,26 +16,20 @@ import no.nav.familie.ba.sak.integrasjoner.organisasjon.OrganisasjonService
 import no.nav.familie.ba.sak.kjerne.autovedtak.fødselshendelse.Resultat
 import no.nav.familie.ba.sak.kjerne.behandling.BehandlingHentOgPersisterService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
-import no.nav.familie.ba.sak.kjerne.brev.DokumentService.Companion.alleredeDistribuertMelding
 import no.nav.familie.ba.sak.kjerne.brev.domene.ManueltBrevRequest
 import no.nav.familie.ba.sak.kjerne.brev.domene.maler.Brevmal
 import no.nav.familie.ba.sak.kjerne.fagsak.Fagsak
 import no.nav.familie.ba.sak.kjerne.fagsak.FagsakRepository
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
-import no.nav.familie.ba.sak.kjerne.steg.BehandlerRolle
 import no.nav.familie.ba.sak.kjerne.steg.grunnlagForNyBehandling.VilkårsvurderingForNyBehandlingService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.AnnenVurderingType
-import no.nav.familie.http.client.RessursException
 import no.nav.familie.kontrakter.felles.BrukerIdType
-import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.familie.kontrakter.felles.arbeidsfordeling.Enhet
 import no.nav.familie.kontrakter.felles.dokarkiv.AvsenderMottaker
 import no.nav.familie.kontrakter.felles.organisasjon.Organisasjon
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import org.springframework.http.HttpStatus
-import org.springframework.web.client.RestClientResponseException
 
 internal class DokumentServiceEnhetstest {
     val integrasjonClient = mockk<IntegrasjonClient>(relaxed = true)
@@ -49,13 +43,8 @@ internal class DokumentServiceEnhetstest {
 
     private val dokumentService: DokumentService = spyk(
         DokumentService(
-            integrasjonClient = integrasjonClient,
-            loggService = mockk(relaxed = true),
-            persongrunnlagService = mockk(relaxed = true),
             journalføringRepository = journalføringRepository,
             taskRepository = mockk(relaxed = true),
-            brevKlient = mockk(relaxed = true),
-            brevService = mockk(relaxed = true),
             vilkårsvurderingService = vilkårsvurderingService,
             vilkårsvurderingForNyBehandlingService = vilkårsvurderingForNyBehandlingService,
             rolleConfig = mockk(relaxed = true),
@@ -63,78 +52,10 @@ internal class DokumentServiceEnhetstest {
             utgåendeJournalføringService = utgåendeJournalføringService,
             fagsakRepository = fagsakRepository,
             organisasjonService = organisasjonService,
-            behandlingHentOgPersisterService = behandlingHentOgPersisterService
+            behandlingHentOgPersisterService = behandlingHentOgPersisterService,
+            dokumentGenereringService = mockk(relaxed = true)
         )
     )
-
-    @Test
-    fun `Skal kalle 'loggBrevIkkeDistribuertUkjentAdresse' ved 400 kode og 'Mottaker har ukjent adresse' melding`() {
-        every { dokumentService.håndterMottakerDødIngenAdressePåBehandling(any(), any(), any()) } returns Unit
-        every {
-            integrasjonClient.distribuerBrev(any(), any())
-        } throws RessursException(
-            httpStatus = HttpStatus.BAD_REQUEST,
-            ressurs = Ressurs.failure(),
-            cause = RestClientResponseException("Mottaker har ukjent adresse", 400, "", null, null, null)
-        )
-
-        val journalpostId = "testId"
-        dokumentService.prøvDistribuerBrevOgLoggHendelse(
-            journalpostId,
-            1L,
-            BehandlerRolle.BESLUTTER,
-            Brevmal.SVARTIDSBREV
-        )
-
-        verify(exactly = 1) { dokumentService.loggBrevIkkeDistribuertUkjentAdresse(any(), any(), any()) }
-    }
-
-    @Test
-    fun `Skal kalle 'håndterMottakerDødIngenAdressePåBehandling' ved 410 Gone svar under distribuering`() {
-        every { dokumentService.håndterMottakerDødIngenAdressePåBehandling(any(), any(), any()) } returns Unit
-        every {
-            integrasjonClient.distribuerBrev(any(), any())
-        } throws RessursException(
-            httpStatus = HttpStatus.GONE,
-            ressurs = Ressurs.failure(),
-            cause = RestClientResponseException("", 410, "", null, null, null)
-        )
-
-        val journalpostId = "testId"
-        dokumentService.prøvDistribuerBrevOgLoggHendelse(
-            journalpostId,
-            1L,
-            BehandlerRolle.BESLUTTER,
-            Brevmal.SVARTIDSBREV
-        )
-
-        verify(exactly = 1) { dokumentService.håndterMottakerDødIngenAdressePåBehandling(any(), any(), any()) }
-    }
-
-    @Test
-    fun `Skal hoppe over distribuering ved 409 Conflict mot dokdist`() {
-        every { dokumentService.logger.info(any()) } returns Unit
-        every { dokumentService.logger.warn(any()) } returns Unit
-
-        every {
-            integrasjonClient.distribuerBrev(any(), any())
-        } throws RessursException(
-            httpStatus = HttpStatus.CONFLICT,
-            ressurs = Ressurs.failure(),
-            cause = RestClientResponseException("", 409, "", null, null, null)
-        )
-
-        val journalpostId = "testId"
-        val behandlingId = 1L
-        dokumentService.prøvDistribuerBrevOgLoggHendelse(
-            journalpostId = journalpostId,
-            behandlingId = behandlingId,
-            loggBehandlerRolle = BehandlerRolle.BESLUTTER,
-            brevmal = Brevmal.SVARTIDSBREV
-        )
-
-        verify { dokumentService.logger.warn(alleredeDistribuertMelding(journalpostId, behandlingId)) }
-    }
 
     @Test
     fun `sendManueltBrev skal journalføre med brukerIdType ORGNR hvis brukers id er 9 siffer, og FNR ellers`() {
